pyproject.toml
[build-system]
requires = [
  "setuptools>=65.6.3",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "scrapy_gdrive_exporter"
version = "0.0.1"

description = "Scrapy feed export storage backend for Google Drive"
requires-python = ">=3.8"

dependencies = [
  "google-api-python-client",
  "google-auth",
  "scrapy"
]

[project.optional-dependencies]
tests = [
  "pytest"
]